+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Changed license from GNU GPL to Apache 2.0
 - Added tool (ache run ReplayCrawl) to replay old crawls using a new configuration file
 - Added near-duplicate page detection using min-hashing and LSH
+- Support ELASTIC format in Kafka data format (issue #155)
 
 ## Version 0.10.0 (2018-01-16)